+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+* Fix wrong size calculations for `Konva.Line` with tension
 * Fix `shape.intersects()` behavior when a node is dragged
 
 ## 7.2.1